requirements.txt
Pillow==5.4.1
Open-Tamil==0.9
